chore: add revenues
