+++ src/containers/Equipments/images.js
@@ -1,0 +1,25 @@
+import A630 from '../../assets/images/equipments/a6-30.png'
+import A860 from '../../assets/images/equipments/a8-60.png'
+import ATP60E from '../../assets/images/equipments/atp60e.png'
+import FTP60E from '../../assets/images/equipments/ftp60e.png'
+import FTS60E from '../../assets/images/equipments/fts60e.png'
+import Lioness from '../../assets/images/equipments/lioness.png'
+import MTS60E from '../../assets/images/equipments/mts60e.png'
+import PrimaPremix from '../../assets/images/equipments/prima-premix.png'
+import M860 from '../../assets/images/equipments/m8-60.png'
+
+const imagesProduct = {
+  A630,
+  A860,
+  ATP60E,
+  FTP60E,
+  FTS60E,
+  Lioness,
+  MTS60E,
+  PrimaPremix,
+  M860,
+}
+
+export {
+  imagesProduct,
+}
