+++ src/containers/Equipments/index.js
@@ -5,27 +5,9 @@ import CompanyLogo from '../../assets/images/logo.svg'
 import { Card } from '../../components'
 import styles from './style.module.css'
 
-import A630 from '../../assets/images/equipments/a6-30.png'
-import A860 from '../../assets/images/equipments/a8-60.png'
-import ATP60E from '../../assets/images/equipments/atp60e.png'
-import FTP60E from '../../assets/images/equipments/ftp60e.png'
-import FTS60E from '../../assets/images/equipments/fts60e.png'
-import Lioness from '../../assets/images/equipments/lioness.png'
-import MTS60E from '../../assets/images/equipments/mts60e.png'
-import PrimaPremix from '../../assets/images/equipments/prima-premix.png'
-import M860 from '../../assets/images/equipments/m8-60.png'
-
-const images = {
-  A630,
-  A860,
-  ATP60E,
-  FTP60E,
-  FTS60E,
-  Lioness,
-  MTS60E,
-  PrimaPremix,
-  M860,
-}
+import {
+  imagesProduct,
+} from './images'
 
 const Equipments = ({
   action,
@@ -48,7 +30,7 @@ const Equipments = ({
           >
             <Card
               action={() => action(id)}
-              image={images[image]}
+              image={imagesProduct[image]}
               title={name}
             />
           </div>
